tests: fs: Make common dirops test
Make test_fs_dirops test common.
